moving global variables to pyro.poutine.runtime
