+++ pyro/poutine/__init__.py
@@ -12,7 +12,7 @@ from .escape_poutine import EscapeMessenger
 from .indep_poutine import IndepMessenger  # noqa: F401
 from .infer_config_poutine import InferConfigMessenger
 from .lift_poutine import LiftMessenger
-from .poutine import _PYRO_STACK, Messenger  # noqa: F401
+from .poutine import Messenger  # noqa: F401
 from .replay_poutine import ReplayMessenger
 from .scale_poutine import ScaleMessenger
 from .trace import Trace  # noqa: F401
