+++ pyro/primitives.py
@@ -14,9 +14,8 @@ import pyro.distributions as dist
 import pyro.infer as infer
 import pyro.poutine as poutine
 from pyro.distributions.distribution import Distribution
-from pyro.params import _MODULE_NAMESPACE_DIVIDER, _PYRO_PARAM_STORE, param_with_module_name
-from pyro.poutine.indep_poutine import _DIM_ALLOCATOR
-from pyro.util import am_i_wrapped, apply_stack, deep_getattr, set_rng_seed  # noqa: F401
+from pyro.poutine.runtime import _DIM_ALLOCATOR, _PYRO_PARAM_STORE, _MODULE_NAMESPACE_DIVIDER, am_i_wrapped, apply_stack
+from pyro.util import deep_getattr, set_rng_seed  # noqa: F401
 
 version_prefix = '0.2.0-a0'
 
@@ -438,3 +437,17 @@ def validation_enabled(is_validate=True):
     finally:
         dist.enable_validation(distribution_validation_status)
         infer.enable_validation(infer_validation_status)
+
+
+def param_with_module_name(pyro_name, param_name):
+    return _MODULE_NAMESPACE_DIVIDER.join([pyro_name, param_name])
+
+
+def module_from_param_with_module_name(param_name):
+    return param_name.split(_MODULE_NAMESPACE_DIVIDER)[0]
+
+
+def user_param_name(param_name):
+    if _MODULE_NAMESPACE_DIVIDER in param_name:
+        return param_name.split(_MODULE_NAMESPACE_DIVIDER)[1]
+    return param_name
